requirements.txt
transformers==4.41.2
sentence-transformers==2.7.0
tokenizers==0.19.1

torch>=2.2.0

spacy==3.7.4
nltk==3.8.1

faiss-cpu==1.8.0
chromadb==0.5.3

numpy<2.0,>=1.24.0
pandas>=2.0.0

# ✅ FIXED FastAPI (no CLI conflict)
fastapi==0.110.0
uvicorn==0.29.0

streamlit==1.35.0

python-dotenv==1.0.1
tqdm==4.66.4
loguru==0.7.2

scikit-learn==1.5.0
rouge-score==0.1.2
sacrebleu==2.4.2

langchain==0.2.5
langchain-community==0.2.5

regex==2024.5.15
pypdf==4.2.0